// src/app/users/interfaces/user.interface.ts
// Generated by https://quicktype.io

export interface User {
    id:           string;
    email:        string;
    firstName:    string;
    lastName:     string;
    phone:        string;
    address:      null;
    typeIdentity: string;
    identity:     string;
    role:         string;
    active:       boolean;
    createAt:     string;
    updateAt:     null;
}

export interface Transaction {
    userId: string;
    value: number;
    typeTransaction: string;
    isAdding: boolean;
    receiptPayment: string
}